+++ src/schema/nexus-typegen.ts
@@ -109,8 +109,8 @@ export interface NexusGenInputs {
   };
   PositionInputField: {
     // input type
-    x: number; // Int!
-    y: number; // Int!
+    x: number; // Float!
+    y: number; // Float!
   };
   UpdateBlueprintInput: {
     // input type
@@ -336,8 +336,8 @@ export interface NexusGenObjects {
   };
   Position: {
     // root type
-    x: number; // Int!
-    y: number; // Int!
+    x: number; // Float!
+    y: number; // Float!
   };
   Query: {};
   ResetConfigPayload: {
@@ -684,8 +684,8 @@ export interface NexusGenFieldTypes {
   };
   Position: {
     // field return type
-    x: number; // Int!
-    y: number; // Int!
+    x: number; // Float!
+    y: number; // Float!
   };
   Query: {
     // field return type
@@ -1039,8 +1039,8 @@ export interface NexusGenFieldTypeNames {
   };
   Position: {
     // field return type name
-    x: 'Int';
-    y: 'Int';
+    x: 'Float';
+    y: 'Float';
   };
   Query: {
     // field return type name
